% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_captures.R
\name{as_captures}
\alias{as_captures}
\title{Make a named vector of capture expressions from one of regular expressions.}
\usage{
as_captures(.regexes, .labels = NULL)
}
\arguments{
\item{.regexes}{\if{html}{\out{<chr>}} a vector of capturing expressions}

\item{.labels}{\if{html}{\out{<chr>}}, optional, names for each capturing expression.}
}
\value{
a vector of expressions, some of which may be named capture groups.
}
\description{
Use blank names for any expressions that you want to match, but not to
capture.
}
\examples{
foo <- c(bar = "\\\\d+", "\\\\D+", baz = "\\\\d+")
as_captures(foo)
}
