% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ASSESSMENT_PATTERNS}
\alias{ASSESSMENT_PATTERNS}
\title{The fields concatenated in a Panorama data field about intervention}
\format{
An object of class \code{character} of length 7.
}
\usage{
ASSESSMENT_PATTERNS
}
\description{
The fields concatenated in a Panorama data field about intervention
}
\seealso{
\code{\link[=wrangle_assessment_fields]{wrangle_assessment_fields()}}
}
\keyword{datasets}
