% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_import_raw_reports.R
\name{batch_import_raw_reports}
\alias{batch_import_raw_reports}
\title{Parse the filenames of and import the data from multiple Panorama csv files}
\usage{
batch_import_raw_reports(.filenames)
}
\arguments{
\item{.filenames}{<chr> a vector of paths to student-level Panorama files}
}
\value{
a table with the fields from \code{\link[=filenames_to_tibble]{filenames_to_tibble()}} and \code{\link[=import_raw_report]{import_raw_report()}}
}
\description{
Parse the filenames of and import the data from multiple Panorama csv files
}
