% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_student_demographics.R
\name{compile_student_demographics}
\alias{compile_student_demographics}
\title{Join tables of individual student demographic data into a single one}
\usage{
compile_student_demographics(.source_table, ...)
}
\arguments{
\item{.source_table}{the table that describes which file a column comes from}

\item{...}{tables of student demographic data as produced by \code{student_info_table()}}
}
\value{
<data_frame>
a data frame with many columns, including source_id, row, School, and
demographic information from each element of \code{...}
}
\description{
Join tables of individual student demographic data into a single one
}
