% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_columns.R
\name{connect_to_columns}
\alias{connect_to_columns}
\title{Join a table of context information to the columns table}
\usage{
connect_to_columns(.frame, .columns_frame)
}
\arguments{
\item{.frame}{a table of context information}

\item{.columns_frame}{the table of column information}
}
\value{
a new data frame joining by \code{column_id}
}
\description{
Join a table of context information to the columns table
}
