% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_skips.R
\name{count_skips}
\alias{count_skips}
\title{Find the number of lines to skip in a text file of data}
\usage{
count_skips(.connection, .header_pattern)
}
\arguments{
\item{.connection}{\verb{<obj>} A connection object, like one gets from \code{\link[=file]{file()}}}

\item{.header_pattern}{\verb{<chr>} A regular expression for \code{\link[=grepl]{grepl()}} to search for}
}
\value{
\verb{<int>} the number of lines searched before the pattern is found
}
\description{
Find the number of lines to skip in a text file of data
}
