% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DEMOGRAPHIC_FIELDS}
\alias{DEMOGRAPHIC_FIELDS}
\title{The fields in a Panorama csv that relate to personal and socioeconomic identity}
\format{
An object of class \code{list} of length 10.
}
\usage{
DEMOGRAPHIC_FIELDS
}
\description{
The fields in a Panorama csv that relate to personal and socioeconomic identity
}
\seealso{
\code{\link[=import_raw_report]{import_raw_report()}} \code{\link[=batch_import_raw_reports]{batch_import_raw_reports()}}
}
\keyword{datasets}
