% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics_from_raw.R
\name{demographics_from_raw}
\alias{demographics_from_raw}
\title{Pull individual student information from a table of raw school data}
\usage{
demographics_from_raw(.raw_data)
}
\arguments{
\item{.raw_data}{an output from running \code{\link[=batch_import_raw_reports]{batch_import_raw_reports()}}}
}
\value{
a table with the following columns
\describe{
\item{School}{\verb{<chr>} the name of the school}
\item{Student Student Number}{\verb{<int>} a unique id for the student}
\item{Student First Name}{\verb{<chr>} the student's given name}
\item{Student Last Name}{\verb{<chr>} the student's family name}
\item{Gender}{\verb{<chr>} "Male" or "Female", so actually sex not gender}
\item{504 Status}{\verb{<chr>} in WI, a 504 plan is a formal, personalized plan for special education }
\item{ELL Status}{\verb{<chr>} whether or not the student is learning English}
\item{Grade Level}{\verb{<chr>} PK, K4, KG, or 1-12}
\item{Date of Birth}{\verb{<date>}}
\item{Race Ethnicity}{\verb{<chr>} federal OMB race/ethnicity}
\item{Special ED Status}{\verb{<chr>} whether or not the student is enrolled in special education}
}
}
\description{
Pull individual student information from a table of raw school data
}
