% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FILE_PATTERNS}
\alias{FILE_PATTERNS}
\title{The fields concatenated in a Panorama csv's name}
\format{
An object of class \code{character} of length 11.
}
\usage{
FILE_PATTERNS
}
\description{
The fields concatenated in a Panorama csv's name
}
\seealso{
\code{\link[=filenames_to_tibble]{filenames_to_tibble()}}
}
\keyword{datasets}
