% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filenames_to_tibble.R
\name{filenames_to_tibble}
\alias{filenames_to_tibble}
\title{Parse Panorama file names into their component fields.}
\usage{
filenames_to_tibble(.filenames, .tz = Sys.timezone())
}
\arguments{
\item{.filenames}{<chr> full paths or basenames to Panorama files.}

\item{.tz}{<db> optional, the timezone the files were downloaded in.}
}
\value{
a tibble with six columns: File, School, Population, Report, Period, and Datetime
}
\description{
Parse Panorama file names into their component fields.
}
\examples{
filenames_to_tibble("StarbuckInternationalSchool_students_ELA_YTD_20241203140402.csv")
}
