% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hercpanorama-package.R
\docType{package}
\name{hercpanorama-package}
\alias{hercpanorama}
\alias{hercpanorama-package}
\title{hercpanorama: Wrangle data from the Panorama Student Information System}
\description{
Student-level data from Panoram is export as wide-formatted csv files. Each file has common identifying and demographic data, then many, possibly unique, columns corresponding to scores on assessments or responses to surveys. The names of these columns include information about the type of data collected as well as when the data were collected. There are also usually two columns for every assessment, a raw score and a grade.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://higherx4racine.github.io/hercpanorama/}
}

}
\author{
\strong{Maintainer}: Ben Taft \email{ben.taft@career2cradle.org}

Other contributors:
\itemize{
  \item Higher Expectations for Racine County [copyright holder]
}

}
\keyword{internal}
