% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_raw_report.R
\name{import_raw_report}
\alias{import_raw_report}
\title{Extract text from a csv-formatted student-level file output from Panorama}
\usage{
import_raw_report(.filename)
}
\arguments{
\item{.filename}{<chr> the full path to the file}
}
\value{
a tibble of student-level data.
}
\description{
The table will have at least 10 columns of demographic and identifying data:
\itemize{
\item \verb{Student Student Number} <int>
\item \verb{Student First Name} <chr>
\item \verb{Student Last Name} <chr>
\item \code{Gender} <chr>
\item \verb{504 Status} <chr>
\item \verb{ELL Status} <chr>
\item \verb{Grade Level} <chr>
\item \verb{Date of Birth} <date>
\item \verb{Race Ethnicity} <chr>
\item \verb{Special ED Status} <chr>
}
}
