% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{INTERVENTION_PATTERNS}
\alias{INTERVENTION_PATTERNS}
\title{The fields concatenated in a Panorama data field about intervention}
\format{
An object of class \code{character} of length 5.
}
\usage{
INTERVENTION_PATTERNS
}
\description{
The fields concatenated in a Panorama data field about intervention
}
\seealso{
\code{\link[=wrangle_intervention_fields]{wrangle_intervention_fields()}}
}
\keyword{datasets}
