% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_wrangling.R
\name{map_wrangling}
\alias{map_wrangling}
\title{Title}
\usage{
map_wrangling(.x, .list_col, .wrangler, ...)
}
\arguments{
\item{.x}{a nested data frame that needs mapped wrangling}

\item{.list_col}{the column of nested data frames}

\item{.wrangler}{the function for wrangling each nested element}

\item{...}{additional columns to include in the output}
}
\value{
a data frame with columns from \code{.wrangler(.list_col)} and \code{...}
}
\description{
Title
}
