% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NWEA_MAP_GROWTH_PATTERNS}
\alias{NWEA_MAP_GROWTH_PATTERNS}
\title{The fields concatenated in a Panorama data ield about MAP growth.}
\format{
An object of class \code{character} of length 8.
}
\usage{
NWEA_MAP_GROWTH_PATTERNS
}
\description{
The fields concatenated in a Panorama data ield about MAP growth.
}
\seealso{
\code{\link[=map_wrangling]{map_wrangling()}}
}
\keyword{datasets}
