% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_panoramel_spreadsheet.R
\name{read_panoramel_spreadsheet}
\alias{read_panoramel_spreadsheet}
\title{Read the pages of a spreadsheet into a list of data frames}
\usage{
read_panoramel_spreadsheet(.filename)
}
\arguments{
\item{.filename}{the full path to the spreadsheet}
}
\value{
a list of data frames, one per sheet in the source file
}
\description{
Read the pages of a spreadsheet into a list of data frames
}
