% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student_info_table.R
\name{student_info_table}
\alias{student_info_table}
\title{Create a one-student, one-field table for a single kind of demographic info}
\usage{
student_info_table(
  .student_table,
  .columns_table,
  .field,
  .measures,
  .value_name
)
}
\arguments{
\item{.student_table}{a context table with student demographic fields}

\item{.columns_table}{the table of individual column information}

\item{.field}{the name of the field to search for in \code{.student_table}}

\item{.measures}{the table of measures where the field's values are stored}

\item{.value_name}{the name of the field in the output table}
}
\value{
a data frame with three columns: source_id, row, and \code{.value_name}
}
\description{
Create a one-student, one-field table for a single kind of demographic info
}
