% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_assessment_fields.R
\name{wrangle_assessment_fields}
\alias{wrangle_assessment_fields}
\title{Tidy assessment data into long format}
\usage{
wrangle_assessment_fields(.panorama_data)
}
\arguments{
\item{.panorama_data}{output from \code{\link[=import_raw_report]{import_raw_report()}}}
}
\value{
a tibble with fields that depend upon the suite of assessments
}
\description{
Tidy assessment data into long format
}
