% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_assessment_scores.R
\name{wrangle_assessment_scores}
\alias{wrangle_assessment_scores}
\title{Condense values that describe student scores into a single table}
\usage{
wrangle_assessment_scores(.panoramel_findings)
}
\arguments{
\item{.panoramel_findings}{a dataframe produced by \code{\link[=read_panoramel_spreadsheet]{read_panoramel_spreadsheet()}}}
}
\value{
a data frame with student scores on specific assessments
}
\description{
Condense values that describe student scores into a single table
}
\seealso{
\code{\link[=connect_to_columns]{connect_to_columns()}}
}
