% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_intervention_fields.R
\name{wrangle_intervention_fields}
\alias{wrangle_intervention_fields}
\title{Title}
\usage{
wrangle_intervention_fields(.panorama_data)
}
\arguments{
\item{.panorama_data}{the result of running \code{\link[=import_raw_report]{import_raw_report()}}}
}
\value{
a table with the following columns
\describe{
\item{Student Student Number}{\verb{<int>} a unique id for the student}
\item{Number}{\verb{<int>} which intervention }
\item{Tier}{\verb{<int>} the intensity of the intervention}
\item{Type and Strategy}{\verb{<chr>}the nature of the intervention}
\item{Status}{\verb{<chr>} what stage the intervention is in}
\item{Start Date}{\verb{<date>}}
}
}
\description{
Title
}
