% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_panoramel_students.R
\name{wrangle_panoramel_students}
\alias{wrangle_panoramel_students}
\title{Condense values that describe student identity into a single table}
\usage{
wrangle_panoramel_students(.panoramel_findings)
}
\arguments{
\item{.panoramel_findings}{a dataframe produced by \code{\link[=read_panoramel_spreadsheet]{read_panoramel_spreadsheet()}}}
}
\value{
a data frame with student demographic information
}
\description{
Condense values that describe student identity into a single table
}
\seealso{
\code{\link[=student_info_table]{student_info_table()}}

\code{\link[=compile_student_demographics]{compile_student_demographics()}}
}
