% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hercpanorama-package.R
\docType{package}
\name{hercpanorama-package}
\alias{hercpanorama}
\alias{hercpanorama-package}
\title{hercpanorama: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://higherx4racine.github.io/hercpanorama/}
}

}
\author{
\strong{Maintainer}: Ben Taft \email{ben.taft@career2cradle.org}

Other contributors:
\itemize{
  \item Higher Expectations for Racine County [copyright holder]
}

}
\keyword{internal}
