% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_raw_report.R
\name{import_raw_report}
\alias{import_raw_report}
\title{Extract text from a csv-formatted student-level file output from Panorama}
\usage{
import_raw_report(.filename)
}
\arguments{
\item{.filename}{\verb{<chr>} the full path to the file}
}
\value{
an object of class \code{tbl_df}, \code{tbl}, \code{data.frame}
with at least 10 columns
\describe{
\item{\verb{Student Student Number}}{ \verb{<int>} the student's unique ID}
\item{\verb{Student First Name}}{ \verb{<chr>} the student's given name}
\item{\verb{Student Last Name}}{ \verb{<chr>} the student's family name}
\item{\code{Gender}}{ \verb{<chr>} female or male}
\item{\verb{504 Status}}{ \verb{<chr>} whether or not the student has a personal learning plan}
\item{\verb{ELL Status}}{ \verb{<chr>} whether or not the student is learning English as an additional language}
\item{\verb{Grade Level}}{ \verb{<chr>} what grade the student is in.}
\item{\verb{Date of Birth}}{ \verb{<date>} students' birthdays don't seem to appear in later data sets}
\item{\verb{Race Ethnicity}}{ \verb{<chr>} one of the OMB '97 race/ethnicites}
\item{\verb{Special ED Status}}{ \verb{<chr>} whethe or not the student participates in special education services}
}
}
\description{
The table will have at least 10 columns of demographic and identifying data:
}
